Move hidden helper commands into functions.
