+++ command/ssh/ssh.go
@@ -81,21 +81,20 @@ $ ssh internal.example.com
 '''`,
 		Subcommands: cli.Commands{
 			certificateCommand(),
+			checkHostCommand(),
 			configCommand(),
-			configHelperCommand(),
-			loginCommand(),
-			logoutCommand(),
+			fingerPrintCommand(),
+			hostsCommand(),
 			inspectCommand(),
 			listCommand(),
-			fingerPrintCommand(),
+			loginCommand(),
+			logoutCommand(),
+			needsRenewalCommand(),
 			// proxyCommand(),
 			proxycommandCommand(),
-			checkHostCommand(),
-			hostsCommand(),
+			rekeyCommand(),
 			renewCommand(),
 			revokeCommand(),
-			rekeyCommand(),
-			needsRenewalCommand(),
 		},
 	}
